Make plugin integration optional

+++ memflow-pcileech/src/lib.rs
@@ -1,6 +1,7 @@
 use parking_lot::Mutex;
 use std::ffi::c_void;
 use std::os::raw::c_char;
+#[cfg(feature = "plugins")]
 use std::path::Path;
 use std::ptr;
 use std::ptr::null_mut;
@@ -10,6 +11,7 @@ use std::sync::Arc;
 use log::LevelFilter;
 use log::{error, info};
 
+#[cfg(feature = "plugins")]
 use memflow::cglue;
 use memflow::mem::phys_mem::*;
 use memflow::prelude::v1::*;
@@ -21,10 +23,11 @@ const PAGE_SIZE: usize = 0x1000usize;
 // the absolute minimum BUF_ALIGN is 4.
 // using 8 bytes as BUF_ALIGN here simplifies things a lot
 // and makes our gap detection code work in cases where page boundaries would be crossed.
-const BUF_ALIGN: usize = 8;
+const BUF_ALIGN: u64 = 8;
 const BUF_MIN_LEN: usize = 8;
 const BUF_LEN_ALIGN: usize = 8;
 
+#[cfg(feature = "plugins")]
 cglue_impl_group!(PciLeech, ConnectorInstance<'a>, {});
 
 fn build_lc_config(device: &str, remote: Option<&str>, with_mem_map: bool) -> LC_CONFIG {
@@ -79,8 +82,8 @@ fn build_lc_config(device: &str, remote: Option<&str>, with_mem_map: bool) -> LC
     }
 }
 
-const fn calc_num_pages(start: usize, size: usize) -> usize {
-    ((start & (PAGE_SIZE - 1)) + size + (PAGE_SIZE - 1)) >> 12
+const fn calc_num_pages(start: u64, size: u64) -> u64 {
+    ((start & (PAGE_SIZE as u64 - 1)) + size + (PAGE_SIZE as u64 - 1)) >> 12
 }
 
 #[allow(clippy::mutex_atomic)]
@@ -100,6 +103,7 @@ impl PciLeech {
         Self::new_internal(device, remote, None, auto_clear)
     }
 
+    #[cfg(feature = "plugins")]
     pub fn with_mem_map_file<P: AsRef<Path>>(
         device: &str,
         remote: Option<&str>,
@@ -226,15 +230,15 @@ impl PhysicalMemory for PciLeech {
         };
 
         // get total number of pages
-        let num_pages = vec.iter().fold(0usize, |acc, read| {
-            acc + calc_num_pages(read.0.to_umem(), read.2.len())
+        let num_pages = vec.iter().fold(0u64, |acc, read| {
+            acc + calc_num_pages(read.0.to_umem(), read.2.len() as u64)
         });
 
         // allocate scatter buffer
         let mut mems = std::ptr::null_mut::<PMEM_SCATTER>();
         let result = unsafe {
             LcAllocScatter2(
-                (num_pages * PAGE_SIZE) as u32,
+                (num_pages * PAGE_SIZE as u64) as u32,
                 std::ptr::null_mut(),
                 num_pages as u32,
                 &mut mems as *mut PPMEM_SCATTER,
@@ -257,7 +261,7 @@ impl PhysicalMemory for PciLeech {
 
                 if addr_align == 0 && len_align == 0 && out.len() >= BUF_MIN_LEN {
                     // properly aligned read
-                    unsafe { (*mem).qwA = page_addr.to_umem() as u64 };
+                    unsafe { (*mem).qwA = page_addr.to_umem() };
                     unsafe { (*mem).__bindgen_anon_1.pb = out.as_mut_ptr() };
                     unsafe { (*mem).cb = out.len() as u32 };
                 } else {
@@ -292,7 +296,7 @@ impl PhysicalMemory for PciLeech {
                         out_end: out.len() + addr_align as usize,
                     });
 
-                    unsafe { (*mem).qwA = page_addr_align as u64 };
+                    unsafe { (*mem).qwA = page_addr_align };
                     unsafe { (*mem).__bindgen_anon_1.pb = buffer_ptr };
                     unsafe { (*mem).cb = buffer_len as u32 };
                 }
@@ -353,15 +357,15 @@ impl PhysicalMemory for PciLeech {
         };
 
         // get total number of pages
-        let num_pages = vec.iter().fold(0usize, |acc, read| {
-            acc + calc_num_pages(read.0.to_umem(), read.2.len())
+        let num_pages = vec.iter().fold(0u64, |acc, read| {
+            acc + calc_num_pages(read.0.to_umem(), read.2.len() as u64)
         });
 
         // allocate scatter buffer
         let mut mems = std::ptr::null_mut::<PMEM_SCATTER>();
         let result = unsafe {
             LcAllocScatter2(
-                (num_pages * PAGE_SIZE) as u32,
+                (num_pages * PAGE_SIZE as u64) as u32,
                 std::ptr::null_mut(),
                 num_pages as u32,
                 &mut mems as *mut PPMEM_SCATTER,
@@ -384,7 +388,7 @@ impl PhysicalMemory for PciLeech {
 
                 if addr_align == 0 && len_align == 0 && out.len() >= BUF_MIN_LEN {
                     // properly aligned write
-                    unsafe { (*mem).qwA = page_addr.to_umem() as u64 };
+                    unsafe { (*mem).qwA = page_addr.to_umem() };
                     unsafe { (*mem).__bindgen_anon_1.pb = out.as_ptr() as *mut u8 };
                     unsafe { (*mem).cb = out.len() as u32 };
                 } else {
@@ -423,7 +427,7 @@ impl PhysicalMemory for PciLeech {
                     });
 
                     // store pointers into pcileech struct for writing (after we dispatched a read)
-                    unsafe { (*mem).qwA = page_addr_align as u64 };
+                    unsafe { (*mem).qwA = page_addr_align };
                     unsafe { (*mem).__bindgen_anon_1.pb = buffer_ptr };
                     unsafe { (*mem).cb = buffer_len as u32 };
                 }
@@ -513,6 +517,7 @@ impl PhysicalMemory for PciLeech {
     }
 }
 
+#[cfg(feature = "plugins")]
 fn validator() -> ArgsValidator {
     ArgsValidator::new()
         .arg(ArgDescriptor::new("default").description("the target device to be used by LeechCore"))
@@ -523,6 +528,7 @@ fn validator() -> ArgsValidator {
 }
 
 /// Creates a new PciLeech Connector instance.
+#[cfg(feature = "plugins")]
 #[connector(name = "pcileech", help_fn = "help", target_list_fn = "target_list")]
 pub fn create_connector(args: &ConnectorArgs) -> Result<PciLeech> {
     let validator = validator();
@@ -557,6 +563,7 @@ pub fn create_connector(args: &ConnectorArgs) -> Result<PciLeech> {
 }
 
 /// Retrieve the help text for the Qemu Procfs Connector.
+#[cfg(feature = "plugins")]
 pub fn help() -> String {
     let validator = validator();
     format!(
@@ -573,6 +580,7 @@ Available arguments are:
 }
 
 /// Retrieve a list of all currently available PciLeech targets.
+#[cfg(feature = "plugins")]
 pub fn target_list() -> Result<Vec<TargetInfo>> {
     // TODO: check if usb is connected, then list 1 target
     Ok(vec![])
